исправил сортировку по зарплате, исправил сохранение superJob, добавил docstr, вывел обстрактные классы отдельно
